Separate out alarm creation per service
