+++ serverless-plugin/cascading-config.js
@@ -20,7 +20,7 @@ function cascade(node, parentNode = {}, depth = 0) {
   const childNodes = {}
   const compiledNode = {}
   for (const [key, value] of Object.entries({ ...parentNode, ...node })) {
-    if (typeof value === 'object') {
+    if (typeof value === 'object' && value !== null) {
       childNodes[key] = value
     } else {
       compiledNode[key] = value
