+++ serverless-test-project/serverless.yml
@@ -7,7 +7,10 @@ plugins:
 custom:
   slicWatch:
     topicArn: {Ref: alarmTopic}
-    invocationsThreshold: 10
+    alarms:
+      Lambda:
+        Invocations:
+          Threshold: 10
 
 provider:
   name: aws
